Move DB recovery to admin page
